Team42 201/Привести импорты в порядок (#277)
* refactor: organize imports across the project

* fix: resolved circular imports in shared folder

* refactor: move text constants to shared/lib/locales

* refactor: simplify imports

* refactor: keep ROUTES import as is
