+++ src/shared/ui/card.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { cn } from '@/shared/lib/utils';
+import { cn } from '../lib/utils';
 
 function Card({ className, ...props }: React.ComponentProps<'div'>) {
   return (
