+++ src/shared/ui/checkbox.tsx
@@ -2,7 +2,7 @@ import { CheckIcon } from 'lucide-react';
 import { Checkbox as CheckboxPrimitive } from 'radix-ui';
 import * as React from 'react';
 
-import { cn } from '@/shared/lib/utils';
+import { cn } from '../lib/utils';
 
 export function Checkbox({
   className,
