+++ src/shared/ui/password-input.tsx
@@ -1,7 +1,8 @@
 import { Eye, EyeOff } from 'lucide-react';
 import * as React from 'react';
 
-import { cn } from '@/shared/lib/utils';
+import { cn } from '../lib/utils';
+
 import { Input } from '@/shared/ui/input';
 
 type PasswordInputProps = React.ComponentProps<'input'>;
